replace fees tooltip on swap
